Performance improvement with bf16 (#438)
* fix for bf16 performance

* fix continue

* fix precision unit test

+++ tests/test_precision_control.py
@@ -13,7 +13,7 @@ pytestmark = pytest.mark.mpi_skip()
 def pytest_resolve_precision_aliases():
     prec, param_dtype, autocast_dtype = tvt.resolve_precision("bfloat16")
     assert prec == "bf16"
-    assert param_dtype == torch.float32
+    assert param_dtype == torch.bfloat16
     assert autocast_dtype == torch.bfloat16
 
     prec, param_dtype, autocast_dtype = tvt.resolve_precision("float32")
